Completly refactored the package
------------------------------------------------------------
First version of driver for LedPulser2

This adds the driver class for the second generation of the
RUB PANDA Light pulser system.
The basic function of the new driver has been verified using
a new light pulser with the most recent firmware.

An open issue is the LCD calibration, i.e. the calculation of a
LCD DAC setting for a given intensity value. The calibration
of the new light pulsers hasn't begun yet, so for now, the
intensity setting just uses a raw DAC value.

Nevertheless, the code is new will require some more testing
before it can be considered stable.
------------------------------------------------------------
Fixed frequency calculation

The frequency calculation was a "magic number formula" and contained
a wrong prescaler value.
The parameters are now given by symbolic constants defined as
static const unsigned (long) int in the header file.
The calculation formula has been adapted accordingly.
------------------------------------------------------------
Fix parameters in LedPulser2

Testing the LEDPulser2 driver in a real environment with EPICS and
CSS showed clearly that using the same asyn parameter for both read
and write access is a rather bad idea. When changing a value in CSS,
the corresponding read value will also instantly change although no
communication on the CAN bus occured. Thus, the user might be fooled
into thinking that his change has already been applied which is not
the case.
To solve this issue, the parameters are split in *_in and *_out
parameters to solve this issue. The "command parameters" read and
write are not affected by this as they do not transport any LED pulser
configuration data.

The frequency is measured in Hertz and not in Kilohertz since the fix
of the calculation in commit 8f5a8db3b3612f98c785e9a1755822ff3fa8dd42,
so its datatype has been changed from epicsFloat64 to epicsInt32 as sub-
Hertz precision is neither needed nor achievable using our micro-
controller.

To avoid code duplication, the code extracting the received data from
a CAN frame and putting it into the correct asyn parameters has been
put in the central function parseReadCanFrame(). It is called from
the read data code, the constructor and also the write data code - as the
new light pulser answers with is current configuration after a config
change command, so the user can see whether the update succeeded.

In a first test in the lab, this version has been proven working at the
Proto192 with a second generation light pulser.
This does however not rule out the possibility of bugs completely.
------------------------------------------------------------
Completly refactored package

The Kernel module for the Raspberry Pi CAN adapter Board is now using socketCAN instead of CHARDEV.
This allows to use one user space program (drvAsynCan) to access all
common CAN bus interfaces from Peak, Kvaser, a.m.m.

Due to the work for the ISEG VDS modules and the ISEG ECH44A crate, the
class drvAsynIsegHv was renamed to drvAsynIsegEhsEds.

Parameter names of asynPortDrivers have been renamed.